fiddling a bit
the new >= 2 thing is super duper slower than previous.

The previous only looked at homegames.

So working on an idea that home games are constrained to a known
value, which in turn means I can just check that homegames are
balanced versus other teams, and assume reflexivity works?  No I don't
think that will work right, as away games are not constrained quite
right.  Although, they will be constrained for the corresponding home
team.  So maybe it will work.